+++ src/index.ts
@@ -11,6 +11,7 @@ export * from './models/request';
 export * from './models/response';
 export * from './models/request-options';
 export * from './models/request-context';
+export * from './models/correlation-id-config';
 
 // Contracts
 export * from './contracts/http-interceptor.contract';